requirements.txt
flask
flask-login
flask-sqlalchemy
pandas
plotly
scikit-learn
joblib